requirements.txt
yt-dlp
ffmpeg-python
google-cloud-speech
google-cloud-storage
flask
nltk
